+++ netreach-v4/val.c
@@ -2,11 +2,11 @@
 #include <sstream>
 #include <arpa/inet.h>
 
-int32_t Val::MAX_VALLEN = 128;
-int32_t Val::SWITCH_MAX_VALLEN = 256;
+uint32_t Val::MAX_VALLEN = 128;
+uint32_t Val::SWITCH_MAX_VALLEN = 256;
 
-int32_t Val::get_padding_size(int32_t vallen) {
-	int32_t padding_size = 0;
+uint32_t Val::get_padding_size(uint32_t vallen) {
+	uint32_t padding_size = 0;
 	if (vallen <= SWITCH_MAX_VALLEN) {
 		if (vallen % 8 != 0) {
 			padding_size = 8 - vallen % 8;
@@ -28,7 +28,7 @@ Val::~Val() {
 	}
 }
 
-Val::Val(const char* buf, int32_t length) {
+Val::Val(const char* buf, uint32_t length) {
 	INVARIANT(buf != nullptr);
 	INVARIANT(length >= 0 && length <= MAX_VALLEN);
 
@@ -143,7 +143,7 @@ bool Val::operator==(const Val &other) {
 		return true;
 	}
 	if (val_length == other.val_length && val_data != nullptr && other.val_data != nullptr) {
-		for (int32_t i = 0; i < val_length; i++) {
+		for (uint32_t i = 0; i < val_length; i++) {
 			if (val_data[i] != other.val_data[i]) {
 				return false;
 			}
@@ -166,7 +166,7 @@ void Val::from_slice(rocksdb::Slice& slice) {
 		INVARIANT(val_data != nullptr);
 		memcpy((char *)val_data, slice.data_, slice.size_);
 
-		val_length = int32_t(slice.size_);
+		val_length = uint32_t(slice.size_);
 	}
 	else {
 		val_length = 0;
@@ -176,14 +176,14 @@ void Val::from_slice(rocksdb::Slice& slice) {
 */
 
 void Val::from_string(std::string& str) {
-	INVARIANT(int32_t(str.length()) <= MAX_VALLEN);
+	INVARIANT(uint32_t(str.length()) <= MAX_VALLEN);
 	if (str.data() != nullptr && str.length() != 0) {
 		// Deep copy
 		val_data = new char[str.length()];
 		INVARIANT(val_data != nullptr);
 		memcpy((char *)val_data, str.data(), str.length());
 
-		val_length = int32_t(str.length());
+		val_length = uint32_t(str.length());
 	}
 	else {
 		val_length = 0;
@@ -193,7 +193,7 @@ void Val::from_string(std::string& str) {
 
 std::string Val::to_string() const { // For print
 	std::stringstream ss;
-	for (int32_t i = 0; i < val_length; i++) {
+	for (uint32_t i = 0; i < val_length; i++) {
 		ss << uint8_t(val_data[i]);
 		if (i != val_length-1) {
 			ss << ":";
@@ -203,7 +203,7 @@ std::string Val::to_string() const { // For print
 }
 
 uint32_t Val::get_bytesnum() const {
-	return uint32_t(val_length);
+	return val_length;
 }
 
 uint32_t Val::deserialize(const char *buf, uint32_t buflen) {
@@ -233,19 +233,19 @@ uint32_t Val::deserialize(const char *buf, uint32_t buflen) volatile {
 uint32_t Val::deserialize_vallen(const char *buf, uint32_t buflen) {
 	INVARIANT(buf != nullptr);
 
-	INVARIANT(buflen >= sizeof(int32_t));
-	val_length = *(const int32_t*)buf; // # of bytes
-	val_length = int32_t(ntohl(uint32_t(val_length))); // Big-endian to little-endian
+	INVARIANT(buflen >= sizeof(uint32_t));
+	val_length = *(const uint32_t*)buf; // # of bytes
+	val_length = ntohl(val_length); // Big-endian to little-endian
 	INVARIANT(val_length >= 0 && val_length <= MAX_VALLEN);
-	return sizeof(int32_t); // sizeof(vallen)
+	return sizeof(uint32_t); // sizeof(vallen)
 }
 
 uint32_t Val::deserialize_val(const char *buf, uint32_t buflen) {
 	INVARIANT(buf != nullptr);
 
-	int32_t padding_size = get_padding_size(val_length); // padding for value <= 128B 
-	int32_t deserialize_size = val_length + padding_size;
-	INVARIANT(deserialize_size >= 0 && buflen >= uint32_t(deserialize_size));
+	uint32_t padding_size = get_padding_size(val_length); // padding for value <= 128B 
+	uint32_t deserialize_size = val_length + padding_size;
+	INVARIANT(deserialize_size >= 0 && buflen >= deserialize_size);
 
 	if (val_data != nullptr) {
 		delete [] val_data;
@@ -261,15 +261,15 @@ uint32_t Val::deserialize_val(const char *buf, uint32_t buflen) {
 uint32_t Val::serialize(char *buf, uint32_t buflen) {
 	INVARIANT(val_data != nullptr);
 
-	int32_t padding_size = get_padding_size(val_length); // padding for value <= 128B 
-	int32_t serialize_size = sizeof(int32_t) + val_length + padding_size;
+	uint32_t padding_size = get_padding_size(val_length); // padding for value <= 128B 
+	uint32_t serialize_size = sizeof(uint32_t) + val_length + padding_size;
 	INVARIANT(serialize_size >= 0 && buflen >= uint32_t(serialize_size));
 
-	int32_t bigendian_vallen = int32_t(htonl(uint32_t(val_length))); // Little-endian to big-endian
-	memcpy(buf, (char *)&bigendian_vallen, sizeof(int32_t)); // Switch needs to use vallen
-	memcpy(buf + sizeof(int32_t), val_data, val_length);
+	uint32_t bigendian_vallen = htonl(val_length); // Little-endian to big-endian
+	memcpy(buf, (char *)&bigendian_vallen, sizeof(uint32_t)); // Switch needs to use vallen
+	memcpy(buf + sizeof(uint32_t), val_data, val_length);
 	if (padding_size > 0) { // vallen <= 128B and vallen % 8 != 0
-		memset(buf + sizeof(int32_t) + val_length, 0, padding_size);
+		memset(buf + sizeof(uint32_t) + val_length, 0, padding_size);
 	}
 	return serialize_size; // sizeof(vallen) + vallen + [padding size]
 }
